Adds preliminaries and proposal part 1
This is the first part of a proposal for an ECMA 402 inspired API
surface in rust.  Incidentally it does not treat any specific part of
ECMA 402, but tries to get a general buy-in for collaborator types that
will be useful in defining ECMA 402.

The API is also inspired by the work of @zbraniecki on `unic-locale`
and `unic-langid`.